requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
djangorestframework==3.10.3
idna==2.8
libmagic==1.0
prompt-toolkit==2.0.10
psycopg2-binary==2.8.4
python-magic==0.4.15
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
wcwidth==0.1.7
wit==5.1.0
